+++ nb_cli/handlers/plugin.py
@@ -8,7 +8,6 @@ from cookiecutter.main import cookiecutter
 from nb_cli.config import ConfigManager
 from nb_cli.prompts import Choice, ListPrompt, InputPrompt, ConfirmPrompt
 
-from ._config import JSONConfig, TOMLConfig
 from ._pip import _call_pip_update, _call_pip_install, _call_pip_uninstall
 from .utils import (
     Plugin,
